requirements.txt
pandas>=1.5.0
openpyxl>=3.1.0
requests>=2.28.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.25.0
groq>=0.4.0
google-generativeai>=0.3.0
streamlit>=1.28.0
argparse
